+++ CHANGELOG.md
@@ -10,12 +10,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `!song request` no longer allows one extra request to enter the queue. ([#33])
 - Avoid playing the same songs over and over ([#35]).
 - Fix off-by-one check in currency transfer (`!currency give <user> <amount>`).
+- Only non-moderator and non-streamer chat bumps the idle counter.
 
 ### Added
 - `!admin version` to check current setmod-bot package version. ([#32])
 - Setting for controlling player feedback in chat (`player/chat-feedback`).
 - Setting for controlling overlay update interval (`player/song-update-interval`).
-- `!admin setting` for reading and writing settings through chat.
+- `!admin settings` for reading and writing settings through chat.
 
 [Unreleased]: https://github.com/udoprog/setmod/compare/0.2.6...HEAD
 [#32]: https://github.com/udoprog/setmod/issues/32